Correct the package name.
