+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/aggregate/JdbcAggregateStorage.java
@@ -29,7 +29,7 @@ import org.spine3.server.entity.Visibility;
 import org.spine3.server.storage.jdbc.DatabaseException;
 import org.spine3.server.storage.jdbc.JdbcStorageFactory;
 import org.spine3.server.storage.jdbc.aggregate.query.AggregateStorageQueryFactory;
-import org.spine3.server.storage.jdbc.entity.status.VisibilityHandler;
+import org.spine3.server.storage.jdbc.entity.visibility.VisibilityHandler;
 import org.spine3.server.storage.jdbc.util.DataSourceWrapper;
 import org.spine3.server.storage.jdbc.util.DbIterator;
