+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/aggregate/query/AggregateStorageQueryFactory.java
@@ -27,12 +27,12 @@ import org.spine3.server.aggregate.AggregateStorage;
 import org.spine3.server.entity.Visibility;
 import org.spine3.server.storage.jdbc.entity.query.InsertAndMarkEntityQuery;
 import org.spine3.server.storage.jdbc.entity.query.MarkEntityQuery;
-import org.spine3.server.storage.jdbc.entity.status.VisibilityHandlingStorageQueryFactory;
-import org.spine3.server.storage.jdbc.entity.status.VisibilityQueryFactories;
-import org.spine3.server.storage.jdbc.entity.status.query.CreateVisibilityTableQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.InsertVisibilityQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.SelectVisibilityQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.UpdateVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.VisibilityHandlingStorageQueryFactory;
+import org.spine3.server.storage.jdbc.entity.visibility.VisibilityQueryFactories;
+import org.spine3.server.storage.jdbc.entity.visibility.query.CreateVisibilityTableQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.InsertVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.SelectVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.UpdateVisibilityQuery;
 import org.spine3.server.storage.jdbc.util.DataSourceWrapper;
 import org.spine3.server.storage.jdbc.util.DbTableNameFactory;
 import org.spine3.server.storage.jdbc.util.IdColumn;
