+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/entity/query/RecordStorageQueryFactory.java
@@ -26,12 +26,12 @@ import org.spine3.server.entity.Entity;
 import org.spine3.server.entity.EntityRecord;
 import org.spine3.server.entity.Visibility;
 import org.spine3.server.storage.RecordStorage;
-import org.spine3.server.storage.jdbc.entity.status.VisibilityHandlingStorageQueryFactory;
-import org.spine3.server.storage.jdbc.entity.status.VisibilityQueryFactories;
-import org.spine3.server.storage.jdbc.entity.status.query.CreateVisibilityTableQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.InsertVisibilityQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.SelectVisibilityQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.UpdateVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.VisibilityHandlingStorageQueryFactory;
+import org.spine3.server.storage.jdbc.entity.visibility.VisibilityQueryFactories;
+import org.spine3.server.storage.jdbc.entity.visibility.query.CreateVisibilityTableQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.InsertVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.SelectVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.UpdateVisibilityQuery;
 import org.spine3.server.storage.jdbc.query.DeleteRowQuery;
 import org.spine3.server.storage.jdbc.util.DataSourceWrapper;
 import org.spine3.server.storage.jdbc.util.DbTableNameFactory;
