+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/entity/visibility/VisibilityHandlingStorageQueryFactory.java
@@ -18,16 +18,16 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.storage.jdbc.entity.status;
+package org.spine3.server.storage.jdbc.entity.visibility;
 
 import org.slf4j.Logger;
 import org.spine3.server.entity.Visibility;
 import org.spine3.server.storage.jdbc.entity.query.InsertAndMarkEntityQuery;
 import org.spine3.server.storage.jdbc.entity.query.MarkEntityQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.CreateVisibilityTableQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.InsertVisibilityQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.SelectVisibilityQuery;
-import org.spine3.server.storage.jdbc.entity.status.query.UpdateVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.CreateVisibilityTableQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.InsertVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.SelectVisibilityQuery;
+import org.spine3.server.storage.jdbc.entity.visibility.query.UpdateVisibilityQuery;
 
 /**
  * And interface for managing a table for storing the {@linkplain Visibility entity visibility}.
