+++ rdbms/src/main/java/org/spine3/server/storage/jdbc/entity/visibility/query/package-info.java
@@ -22,6 +22,6 @@
  * @author Dmytro Dashenkov.
  */
 @ParametersAreNonnullByDefault
-package org.spine3.server.storage.jdbc.entity.status;
+package org.spine3.server.storage.jdbc.entity.visibility.query;
 
 import javax.annotation.ParametersAreNonnullByDefault;
